+++ app/src/main/java/org/simple/clinic/enterotp/EnterOtpScreen.kt
@@ -1,34 +1,39 @@
 package org.simple.clinic.enterotp
 
 import android.content.Context
-import android.os.Parcelable
-import android.util.AttributeSet
+import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
-import android.widget.RelativeLayout
 import androidx.transition.TransitionManager
 import com.jakewharton.rxbinding3.view.clicks
 import com.jakewharton.rxbinding3.widget.editorActions
+import com.spotify.mobius.functions.Consumer
 import io.reactivex.Observable
-import io.reactivex.rxkotlin.ofType
+import io.reactivex.rxkotlin.cast
+import kotlinx.parcelize.Parcelize
 import org.simple.clinic.LOGIN_OTP_LENGTH
 import org.simple.clinic.R
 import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.appconfig.Country
 import org.simple.clinic.databinding.ScreenEnterotpBinding
 import org.simple.clinic.di.injector
-import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.navigation.v2.Router
-import org.simple.clinic.util.unsafeLazy
+import org.simple.clinic.navigation.v2.ScreenKey
+import org.simple.clinic.navigation.v2.fragments.BaseScreen
 import org.simple.clinic.widgets.UiEvent
 import org.simple.clinic.widgets.hideKeyboard
 import org.simple.clinic.widgets.showKeyboard
 import javax.inject.Inject
 
-class EnterOtpScreen(
-    context: Context,
-    attributeSet: AttributeSet
-) : RelativeLayout(context, attributeSet), EnterOtpUi, EnterOtpUiActions {
+class EnterOtpScreen : BaseScreen<
+    EnterOtpScreen.Key,
+    ScreenEnterotpBinding,
+    EnterOtpModel,
+    EnterOtpEvent,
+    EnterOtpEffect,
+    Unit>(), EnterOtpUi, EnterOtpUiActions {
 
   @Inject
   lateinit var router: Router
@@ -39,85 +44,67 @@ class EnterOtpScreen(
   @Inject
   lateinit var effectHandlerFactory: EnterOtpEffectHandler.Factory
 
-  private var binding: ScreenEnterotpBinding? = null
-
   private val otpEntryEditText
-    get() = binding!!.otpEntryEditText
+    get() = binding.otpEntryEditText
 
   private val backButton
-    get() = binding!!.backButton
+    get() = binding.backButton
 
   private val resendSmsButton
-    get() = binding!!.resendSmsButton
+    get() = binding.resendSmsButton
 
   private val userPhoneNumberTextView
-    get() = binding!!.userPhoneNumberTextView
+    get() = binding.userPhoneNumberTextView
 
   private val smsSentTextView
-    get() = binding!!.smsSentTextView
+    get() = binding.smsSentTextView
 
   private val errorTextView
-    get() = binding!!.errorTextView
+    get() = binding.errorTextView
 
   private val validateOtpProgressBar
-    get() = binding!!.validateOtpProgressBar
+    get() = binding.validateOtpProgressBar
 
   private val otpEntryContainer
-    get() = binding!!.otpEntryContainer
+    get() = binding.otpEntryContainer
 
-  private val events by unsafeLazy {
-    Observable
-        .mergeArray(
-            otpSubmits(),
-            resendSmsClicks()
-        )
-        .compose(ReportAnalyticsEvents())
-  }
+  private val rootLayout
+    get() = binding.rootLayout
 
-  private val delegate by unsafeLazy {
-    val uiRenderer = EnterOtpUiRenderer(this)
+  override fun defaultModel() = EnterOtpModel.create()
 
-    MobiusDelegate.forView(
-        events = events.ofType(),
-        defaultModel = EnterOtpModel.create(),
-        update = EnterOtpUpdate(LOGIN_OTP_LENGTH),
-        effectHandler = effectHandlerFactory.create(this).build(),
-        init = EnterOtpInit(),
-        modelUpdateListener = uiRenderer::render
-    )
-  }
+  override fun bindView(
+      layoutInflater: LayoutInflater,
+      container: ViewGroup?
+  ) = ScreenEnterotpBinding.inflate(layoutInflater, container, false)
 
-  override fun onFinishInflate() {
-    super.onFinishInflate()
-    if (isInEditMode) {
-      return
-    }
+  override fun events() = Observable
+      .mergeArray(
+          otpSubmits(),
+          resendSmsClicks()
+      )
+      .compose(ReportAnalyticsEvents())
+      .cast<EnterOtpEvent>()
 
-    binding = ScreenEnterotpBinding.bind(this)
+  override fun createUpdate() = EnterOtpUpdate(LOGIN_OTP_LENGTH)
 
+  override fun createEffectHandler(
+      viewEffectsConsumer: Consumer<Unit>
+  ) = effectHandlerFactory.create(this).build()
+
+  override fun createInit() = EnterOtpInit()
+
+  override fun uiRenderer() = EnterOtpUiRenderer(this)
+
+  override fun onAttach(context: Context) {
+    super.onAttach(context)
     context.injector<Injector>().inject(this)
+  }
 
+  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
     otpEntryEditText.showKeyboard()
     backButton.setOnClickListener { goBack() }
-  }
-
-  override fun onAttachedToWindow() {
-    super.onAttachedToWindow()
-    delegate.start()
-  }
-
-  override fun onDetachedFromWindow() {
-    delegate.stop()
-    binding = null
-    super.onDetachedFromWindow()
-  }
-
-  override fun onSaveInstanceState(): Parcelable? {
-    return delegate.onSaveInstanceState(super.onSaveInstanceState())
-  }
-
-  override fun onRestoreInstanceState(state: Parcelable?) {
-    super.onRestoreInstanceState(delegate.onRestoreInstanceState(state))
   }
 
   private fun otpSubmits(): Observable<UiEvent> {
@@ -147,7 +134,7 @@ class EnterOtpScreen(
   }
 
   override fun goBack() {
-    hideKeyboard()
+    rootLayout.hideKeyboard()
     router.pop()
   }
 
@@ -180,13 +167,13 @@ class EnterOtpScreen(
   }
 
   override fun showProgress() {
-    TransitionManager.beginDelayedTransition(this)
+    TransitionManager.beginDelayedTransition(rootLayout)
     validateOtpProgressBar.visibility = View.VISIBLE
     otpEntryContainer.visibility = View.INVISIBLE
   }
 
   override fun hideProgress() {
-    TransitionManager.beginDelayedTransition(this)
+    TransitionManager.beginDelayedTransition(rootLayout)
     validateOtpProgressBar.visibility = View.INVISIBLE
     otpEntryContainer.visibility = View.VISIBLE
   }
@@ -201,5 +188,12 @@ class EnterOtpScreen(
 
   interface Injector {
     fun inject(target: EnterOtpScreen)
+  }
+
+  @Parcelize
+  data class Key(
+      override val analyticsName: String = "Enter Login OTP Manually"
+  ) : ScreenKey() {
+    override fun instantiateFragment() = EnterOtpScreen()
   }
 }