+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Change large component corner shape to 8dp
 - Add the `CallResult` database table
 - Add 32dp horizontal margin for `noPatientsInFacilityTextView` in `screen_instant_search`
+- Migrate `EnterOtpScreen` to a Fragment
 
 ### Changes
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen